backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mediapipe==0.10.9
insightface==0.7.3
onnxruntime==1.16.3
faiss-cpu==1.7.4
numpy==1.24.3
opencv-python==4.9.0.80
pillow==10.2.0
sqlalchemy==2.0.25
cryptography==42.0.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
